fix: part 3 apparently vscode renames with only capitalizations are not recognized by git
