+++ hierarchical-image-explorer/src/components/SingleImageDisplay.svelte
